// src/Pages/Detail/type.ts
export type NewbieDataType = {
  admin_id: number;
  profile: string;
  name: string;
  phone_number: string;
  address: string;
  first_visit: string;
  birth_date: string;
  is_baptized: boolean;
  gender: string;
  guide: string;
  job: string;
  responsibility: string;
  description: string;
  orientation: string;
  first_class: string;
  second_class: string;
  third_class: string;
  fourth_class: string;
  settled_date: string;
};

export type TeamDataType = {
  team_id: number;
  team_name: string;
  name: string;
};